refactor: 스낵바 렌더함수에서 불필요한 매개변수 제거
