+++ src/js/utils/DOM.js
@@ -3,10 +3,10 @@ export const $$ = (selector) => document.querySelectorAll(selector);
 
 export const isEndOfScroll = ($element) => $element.scrollHeight - $element.scrollTop === $element.clientHeight;
 
-export const showSnackbar = ({ messenger, message, showtime }) => {
-  messenger.innerText = message;
-  messenger.classList.add('show');
-  setTimeout(() => {
-    messenger.classList.remove('show');
-  }, showtime);
-};
+// export const showSnackbar = ({ messenger, message, showtime }) => {
+//   messenger.innerText = message;
+//   messenger.classList.add('show');
+//   setTimeout(() => {
+//     messenger.classList.remove('show');
+//   }, showtime);
+// };
