+++ src/js/utils/snackbar.js
@@ -1,0 +1,11 @@
+import { $ } from './DOM.js';
+
+const $snackbar = $('.snackbar');
+
+export const showSnackbar = ({ message, showtime }) => {
+  $snackbar.innerText = message;
+  $snackbar.classList.add('show');
+  setTimeout(() => {
+    $snackbar.classList.remove('show');
+  }, showtime);
+};
